Refactoring of escape wrapper.
