+++ src/proteus/escape/wrapper_escape.py
@@ -7,13 +7,13 @@ import mors
 import numpy as np
 
 from proteus.utils.constants import element_list, ergcm2stoWm2, secs_per_year
-from proteus.utils.zephyrus import EL_escape
+from zephyrus.escape import EL_escape
 
 log = logging.getLogger("fwl."+__name__)
 
 # Define global variables
-Fxuv_star_SI_full   = None
-age_star_mors       = None
+Fxuv_star_SI_full = None
+age_star_mors = None
 
 
 def RunDummyEsc(hf_row:dict, dt:float, phi_bulk:float):
@@ -36,33 +36,7 @@ def RunDummyEsc(hf_row:dict, dt:float, phi_bulk:float):
     """
     log.info("Running dummy escape...")
 
-    # store value
-    out = {}
-    out["rate_bulk"] = phi_bulk
-
-    # calculate total mass of volatiles (except oxygen, which is set by fO2)
-    M_vols = 0.0
-    for e in element_list:
-        if e=='O':
-            continue
-        M_vols += hf_row[e+"_kg_total"]
-
-
-    # for each elem, calculate new total inventory while
-    # maintaining a constant mass mixing ratio
-    for e in element_list:
-        if e=='O':
-            continue
-
-        # current elemental mass ratio in total
-        emr = hf_row[e+"_kg_total"]/M_vols
-
-        log.debug("    %s mass ratio = %.2e "%(e,emr))
-
-        # new total mass of element e, keeping a constant mixing ratio of that element
-        out[e+"_kg_total"] = emr * (M_vols - phi_bulk * dt * secs_per_year)
-
-    return out
+    return SpeciesEscapeFromTotalEscape(hf_row,dt,phi_bulk)
 
 
 def RunZEPHYRUS(hf_row, dt, M_star,Omega_star,tidal_contribution, semi_major_axis, eccentricity, M_planet, epsilon, R_earth, Rxuv):
@@ -103,30 +77,41 @@ def RunZEPHYRUS(hf_row, dt, M_star,Omega_star,tidal_contribution, semi_major_axi
 
     log.info("Running EL escape (ZEPHYRUS) ...")
 
-    ## Step 1 : Load stellar evolution track + compute EL escape
+    ## Load stellar evolution track + compute EL escape
     log.info("Step 1 : Load stellar evolution track + compute EL escape ")
 
     # Get the age of the star at time t to compute XUV flux at that time
-    age_star        = hf_row["age_star"]                                                # [years]
+    age_star = hf_row["age_star"] # [years]
 
     if (Fxuv_star_SI_full is None):
-        star                = mors.Star(Mstar=M_star, Age=age_star/1e6, Omega=Omega_star)
-        age_star_mors       = star.Tracks['Age']
-        Fxuv_star_SI_full   = ((star.Tracks['Lx'] + star.Tracks['Leuv']) / (4 * np.pi * (semi_major_axis * 1e2)**2)) * ergcm2stoWm2
+        star = mors.Star(Mstar=M_star, Age=age_star/1e6, Omega=Omega_star)
+        age_star_mors = star.Tracks['Age']
+        Fxuv_star_SI_full = ((star.Tracks['Lx'] + star.Tracks['Leuv']) / (4 * np.pi * (semi_major_axis * 1e2)**2)) * ergcm2stoWm2
 
     # Interpolating the XUV flux at the age of the star
-    Fxuv_star_SI        = np.interp(age_star, age_star_mors * 1e6, Fxuv_star_SI_full)                                            # Interpolate to get Fxuv at age_star
+    Fxuv_star_SI = np.interp(age_star, age_star_mors * 1e6, Fxuv_star_SI_full) # Interpolate to get Fxuv at age_star
     log.info(f"Interpolated Fxuv_star_SI at age_star = {age_star} years is {Fxuv_star_SI}")
 
     # Compute energy-limited escape
-    mlr                 = EL_escape(tidal_contribution, semi_major_axis, eccentricity, M_planet, M_star, epsilon, R_earth, Rxuv, Fxuv_star_SI)   # [kg s-1]
+    mlr = EL_escape(tidal_contribution,
+                    semi_major_axis,
+                    eccentricity,
+                    M_planet,
+                    M_star,
+                    epsilon,
+                    R_earth,
+                    Rxuv,
+                    Fxuv_star_SI)   # [kg s-1]
 
-    ## Step 2 : Updated total elemental mass inventories
-    log.info("Step 2 : Updated total elemental mass inventories")
+    log.info('Zephyrus escape computation done :)')
+
+    return SpeciesEscapeFromTotalEscape(hf_row,dt,mlr)
+
+def SpeciesEscapeFromTotalEscape(hf_row:dict, dt:float, esc:float)
 
     # store mass loss rate value
     out = {}
-    out["rate_bulk"] = mlr
+    out["rate_bulk"] = esc
 
     # calculate total mass of volatiles (except oxygen, which is set by fO2)
     M_vols = 0.0
@@ -148,8 +133,6 @@ def RunZEPHYRUS(hf_row, dt, M_star,Omega_star,tidal_contribution, semi_major_axi
         log.debug("    %s mass ratio = %.2e "%(e,emr))
 
         # new total mass of element e, keeping a constant mixing ratio of that element
-        out[e+"_kg_total"] = emr * (M_vols - mlr * dt * secs_per_year)
-
-    log.info('Zephyrus escape computation done :)')
+        out[e+"_kg_total"] = emr * (M_vols - esc * dt * secs_per_year)
 
     return out
